Tungsten.Structure.Graph

Description

This module defines a type isomorphic to algebraic graphs (from A. Mokhov, see the algebraic-graphs package), in terms of Fix from Tungsten.Fix.

A good consumer is a function that can be fused with a good producer. A good producer is a function that can be fused with a good consumer.

Algebraic graphs as fixed-points

data GraphF a b

The "factored-out" recursive type for algebraic graphs.

Constructors

EmptyF

VertexF a

OverlayF b b

ConnectF b b

type Graph a = Fix (GraphF a)

empty :: Graph a

The empty graph.

vertex :: a -> Graph a

A vertex.

overlay :: Graph a -> Graph a -> Graph a

Overlay two graphs.

connect :: Graph a -> Graph a -> Graph a

Connect two graphs.

Classical operations on graphs

foldg

Arguments

:: b
Empty case

-> (a -> b)
Vertex case

-> (b -> b -> b)
Overlay case

-> (b -> b -> b)
Connect case

-> Graph a
The graph to fold

-> b

Fold a graph. Good consumer.

mapg :: (a -> b) -> Graph a -> Graph b

fmap for algebraic graphs. Good consumer and good producer.

bind :: Graph a -> (a -> Graph b) -> Graph b

bind for algebraic graphs. Good consumer and good producer.

Operations on graphs

hasVertex :: Eq a => a -> Graph a -> Bool

Test if a vertex is in a graph. Good consumer.

Conversions

edges :: [(a, a)] -> Graph a

Construct a graph from a list of edges Good consumer of lists and producer of graphs.