Tungsten.Structure.List

Description

This module defines a type isomorphic to linked lists, in terms of Fix from Tungsten.Fix.

A good consumer is a function that can be fused with a good producer. A good producer is a function that can be fused with a good consumer.

Lists as fixed-points

data ListF a b

The factored-out recursive type for lists.

Constructors

NilF

ConsF a b

type List a = Fix (ListF a)

Linked lists as a fixed-point.

nil :: List a

The empty list. Similar to 'Prelude.[]' for Prelude lists.

cons :: a -> List a -> List a

The cons operator. Similar to 'Prelude.(:)' for Prelude lists.

Classical operations on lists

foldr :: (a -> b -> b) -> b -> List a -> b

The classical right fold. Good consumer.

map :: (a -> b) -> List a -> List b

The classical map. Good consumer and good producer.

Operations on lists

elem :: Eq a => a -> List a -> Bool

Search an element in a list. Good consumer.

range :: Int -> Int -> List Int

range start end will produce a list containing int in ascending order from start (inclusive) to end (exclusive). Good producer.

Conversions

toList :: List a -> [a]

Transform a fixed-point list into a Prelude one. Good producer (of Prelude lists) and good consumer (of fixed-point lists).

fromList :: [a] -> List a

Transform a Prelude list into a fixed-point one. Good producer (fixed-point lists) and good consumer of (of Prelude lists).