Tungsten.Structure.Tree

Description

This module defines a type isomorphic to binary trees, in terms of Fix from Tungsten.Fix.

A good consumer is a function that can be fused with a good producer. A good producer is a function that can be fused with a good consumer.

Binary trees as fixed-point

data TreeF a b

The "factored-out" recursive type for binary trees.

Constructors

EmptyF

LeafF a

NodeF b b

type Tree a = Fix (TreeF a)

Binary trees expressed as a fixed-point.

empty :: Tree a

The empty tree.

leaf :: a -> Tree a

A leaf.

node :: Tree a -> Tree a -> Tree a

A node.

Classical operations on trees

mapt :: (a -> b) -> Tree a -> Tree b

fmap for trees. Good consumer and good producer.

bind :: Tree a -> (a -> Tree b) -> Tree b

bind for trees. Good consumer and good producer.

Operations on trees

hasLeaf :: Eq a => a -> Tree a -> Bool

hasLeaf s t tests if the leaf s is present in the tree t. Good consumer.

treeFromList :: [Tree a] -> Tree a

Construct a binary tree from a list. Good consumer (of Prelude lists) and good producer (of trees).

leftTreeN :: Int -> Tree Int

leftTree n construct a tree with n leaves from 1 to n. Good producer.